+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - badge with '+ additional titles' if there is more than one title present
 - allow to toggle all primary sources back on within the edit page
+- ingest button of already imported items are always deactivated
 
 ### Changes